backend-python/requirements.txt
APScheduler==3.8.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
cycler==0.11.0
Cython==0.29.23
distlib==0.3.3
filelock==3.1.0
Flask==2.0.1
Flask-Classful==0.14.2
gensim==4.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.1.0
jsonpickle==2.0.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
msvc-runtime==14.29.30036
numpy==1.21.2
Pillow==8.4.0
pipenv==2021.5.29
platformdirs==2.3.0
psutil==5.8.0
psycopg2-binary==2.9.1
pyparsing==3.0.5
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
sqlite-bro==0.9.3
threadpoolctl==3.0.0
tzdata==2021.5
tzlocal==4.1
virtualenv==20.8.1
virtualenv-clone==0.5.7
waitress==2.0.0
Werkzeug==2.0.1
winpython==4.3.20210620
